Converted to ES6 modules
Initially the project used plain .ts files in global scope, and the TS `outFile` setting to compile into a single output .js, but this quickly lead to issues with file order.

One solution is to use `<reference ...>` tags at the top of every file, indicating what types/objects/etc are used, which TSC uses as a hint for ordering files during compilation. That's a lot of work and a huge pain, so the project was switched to modules instead, for flexibility and ease of use.

Using ES6 modules with p5.js simply means any global functions picked up by the sketch must be explicitly defined on the global object, ex. `window.setup = ` or `globalThis.setup = `. It also allows the use of source maps (so that TS appears in the browser dev console instead of JS), and removes the need to disable eslint `no-unused-vars` errors on every p5 global function.